global: simplified api and packaging
* INCOMPATIBLE: this breaks the backwards compatibility.
