+++ tests/unit/test_merger_arxiv2arxiv.py
@@ -33,8 +33,8 @@ from mock import MagicMock
 import pytest
 
 from json_merger.conflict import Conflict
-from inspire_json_merger import inspire_json_merger
-from inspire_json_merger.inspire_json_merger import inspire_json_merge
+from inspire_json_merger import api
+from inspire_json_merger.api import merge
 from inspire_schemas.api import load_schema
 
 from utils import assert_ordered_conflicts, validate_subschema
@@ -42,10 +42,10 @@ from utils import assert_ordered_conflicts, validate_subschema
 
 @pytest.fixture(autouse=True, scope='module')
 def mock_get_acquisition_source():
-    original_func = inspire_json_merger.get_acquisition_source
-    inspire_json_merger.get_acquisition_source = MagicMock(return_value='arxiv')
+    original_func = api.get_acquisition_source
+    api.get_acquisition_source = MagicMock(return_value='arxiv')
     yield
-    inspire_json_merger.get_acquisition_source = original_func
+    api.get_acquisition_source = original_func
 
 
 # list of all the schema's keys tested
@@ -77,7 +77,7 @@ def test_merging_collections_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -142,7 +142,7 @@ def test_merging_desy_bookkeeping_field():
         Conflict('SET_FIELD', ('_desy_bookkeeping', 0, 'status'), 'abs')
     ]
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -170,7 +170,7 @@ def test_merging_export_to_field():
         Conflict('SET_FIELD', ('_export_to', 'CDS'), False)
     ]
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -222,7 +222,7 @@ def test_merging_private_notes_field():
         )
     ]
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -262,7 +262,7 @@ def test_merging_abstracts_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -304,7 +304,7 @@ def test_merging_accelerator_experiments_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -330,7 +330,7 @@ def test_merging_acquisition_source_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -389,7 +389,7 @@ def test_merging_arxiv_eprints_field():
     }
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -433,7 +433,7 @@ def test_merging_authors_field():
 
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -494,7 +494,7 @@ def test_merging_affiliations_field_per_ref():
 
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -555,7 +555,7 @@ def test_merging_affiliations_field_per_value():
 
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -608,7 +608,7 @@ def test_merging_alternative_names_field():
     }
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -656,7 +656,7 @@ def test_merging_credit_roles_field():
     }
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -686,7 +686,7 @@ def test_merging_curated_relation_field():
 
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -723,7 +723,7 @@ def test_merging_emails_field():
 
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -757,7 +757,7 @@ def test_merging_full_name_field():
 
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -796,7 +796,7 @@ def test_merging_ids_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -829,7 +829,7 @@ def test_merging_inspire_roles_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -868,7 +868,7 @@ def test_merging_raw_affiliations_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -901,7 +901,7 @@ def test_merging_record_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -930,7 +930,7 @@ def test_merging_signature_block_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -959,7 +959,7 @@ def test_merging_uuid_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1007,7 +1007,7 @@ def test_merging_book_series_field():
         )
     ]
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1022,7 +1022,7 @@ def test_merging_citeable_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1082,7 +1082,7 @@ def test_merging_collaborations_field():
         Conflict('SET_FIELD', ('collaborations', 0, 'value'), 'ATLAS')
     ]
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1098,7 +1098,7 @@ def test_merging_control_number_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1143,7 +1143,7 @@ def test_merging_copyright_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1158,7 +1158,7 @@ def test_merging_core_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1187,7 +1187,7 @@ def test_merging_corporate_author_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1202,7 +1202,7 @@ def test_merging_deleted_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1237,7 +1237,7 @@ def test_merging_deleted_records_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1252,7 +1252,7 @@ def test_merging_document_type():
     expected_merged = update  # since the list rule is 'Keep update'
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1297,7 +1297,7 @@ def test_merging_dois_field():
     }
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1312,7 +1312,7 @@ def test_merging_editions_field():
     expected_merged = {'editions': ['editionA', 'edition2']}
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1327,7 +1327,7 @@ def test_merging_energy_ranges_field():
     expected_merged = update  # just update the record with newcoming info
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1374,7 +1374,7 @@ def test_merging_external_system_identifiers_field():
     # so it appears in the conflicts
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1413,7 +1413,7 @@ def test_merging_funding_info_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1455,7 +1455,7 @@ def test_merging_imprints_field():
     # so the head is directly removed, loosing eventually curated info
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1505,7 +1505,7 @@ def test_merging_inspire_categories_field():
     ]}
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1545,7 +1545,7 @@ def test_merging_isbns_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1589,7 +1589,7 @@ def test_merging_keywords_field():
     ]}
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1605,7 +1605,7 @@ def test_merging_languages_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1620,7 +1620,7 @@ def test_merging_legacy_creation_date_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1671,7 +1671,7 @@ def test_merging_license_field():
     ]}
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1686,7 +1686,7 @@ def test_merging_new_record_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1701,7 +1701,7 @@ def test_merging_new_record_field_filled_root():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1716,7 +1716,7 @@ def test_merging_number_of_pages_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1749,7 +1749,7 @@ def test_merging_persistent_identifiers_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1764,7 +1764,7 @@ def test_merging_preprint_date_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1784,7 +1784,7 @@ def test_merging_public_notes_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1861,7 +1861,7 @@ def test_merging_publication_info_field():
     }
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1876,7 +1876,7 @@ def test_merging_publication_type_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1891,7 +1891,7 @@ def test_merging_refereed_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1926,7 +1926,7 @@ def test_merging_report_numbers_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -1941,7 +1941,7 @@ def test_merging_self_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
 
@@ -1955,7 +1955,7 @@ def test_merging_texkeys_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2013,7 +2013,7 @@ def test_merging_thesis_info_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2052,7 +2052,7 @@ def test_merging_title_translations_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2098,7 +2098,7 @@ def test_merging_titles_field():
     ]}
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2119,7 +2119,7 @@ def test_merging_urls_field():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2134,7 +2134,7 @@ def test_merging_wirthdrawn_field():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2174,7 +2174,7 @@ def test_merging_references_field_curated_relation():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2225,7 +2225,7 @@ def test_merging_references_field_raw_refs():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2316,7 +2316,7 @@ def test_merging_references_field_reference_authors():
         ]
     }
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2362,7 +2362,7 @@ def test_merging_references_field_reference_arxiv_eprint():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2430,7 +2430,7 @@ def test_merging_references_field_reference_book_series():
         ]
     }
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2467,7 +2467,7 @@ def test_merging_references_field_reference_document_type():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2504,7 +2504,7 @@ def test_merging_references_field_reference_imprint():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2541,7 +2541,7 @@ def test_merging_references_field_reference_isbn():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2578,7 +2578,7 @@ def test_merging_references_field_reference_label():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2619,7 +2619,7 @@ def test_merging_references_field_reference_misc():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2666,7 +2666,7 @@ def test_merging_references_field_reference_persistent_identifiers():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2728,7 +2728,7 @@ def test_merging_references_field_reference_report_numbers():
         ]
     }
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2765,7 +2765,7 @@ def test_merging_references_field_reference_texkey():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2808,7 +2808,7 @@ def test_merging_references_field_reference_title():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2863,7 +2863,7 @@ def test_merging_references_field_reference_urls():
     expected_conflict = []
     expected_merged = update
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2886,7 +2886,7 @@ def test_self_field():
 
     expected_merged = head
     expected_conflict = []
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -2925,7 +2925,7 @@ def test_record_affiliations_field():
             update['record_affiliations'][0]
         ]
     }
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert conflict == []
     validate_subschema(merged)
@@ -2958,7 +2958,7 @@ def test_related_records_field():
             update['related_records'][0]
         ]
     }
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert conflict == []
     validate_subschema(merged)
@@ -2976,7 +2976,7 @@ def test_curated():
     expected_merged = head
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update, head_source='arxiv')
+    merged, conflict = merge(root, head, update, head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -3021,8 +3021,8 @@ def test_figures():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update,
-                                          head_source='arxiv')
+    merged, conflict = merge(root, head, update,
+                             head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
@@ -3068,8 +3068,8 @@ def test_documents():
     expected_merged = update
     expected_conflict = []
 
-    merged, conflict = inspire_json_merge(root, head, update,
-                                          head_source='arxiv')
+    merged, conflict = merge(root, head, update,
+                             head_source='arxiv')
     assert merged == expected_merged
     assert_ordered_conflicts(conflict, expected_conflict)
     validate_subschema(merged)
